requirements.txt
bcrypt==3.2.2
cffi==1.15.0
colorama==0.4.5
cryptography==37.0.2
future==0.18.2
netmiko==4.1.1
ntc-templates==3.0.0
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0
scp==0.14.4
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
